Weapons!
- Instead of SpaceShip being responsible for firing Bullets, there is
now an abstract class Weapon (a subclass of which is BulletWeapon) that
deals with that. This means that a Spaceship can have multiple weapons
to switch between (this will likely be implemented in the next commit).

- player's main weapon is now a three-shot gun

- new Controllable interface, which SpaceShip implements (Very likely to
change in the future), which takes handleKeyPress() and handleKeysHeld()
out of GameObject. They felt very unnatural there.

- SpaceShip now has a die() method, with the same functionality as
before
- except now the SpaceShips don't behave as if they have a tactical nuke
on board, and their explosions deal no damage.